+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - **Other Features**
   - Added `Spiral\Scaffolder\Command\InfoCommand` console command for getting information about available scaffolder 
     commands.
+  - [spiral/core] Added the ability to bind the interface as a proxy via `Spiral\Core\Config\Proxy` or `Spiral\Core\Config\DeprecationProxy`.
 
 ## 3.11.1 - 2023-12-29